[JRO] Workaround for standalone ios status bar
If you install this app in your home screen as a standalone app it will
open with a nice translucent status bar hovering over the app. While
that's good, it will hover over the main navigation bar at top. Because
of how I built that area for mobile devices this was troublesome.

Using modernizr we now detect whether we're in standalone mode and
adjust the styles of the breadcrumbs and site nav to work with the
hovering status bar.

Also adding an apple-touch-icon-precomposed.png to app-root to sate the
neverending requests for that file from random ios devices, despite what
I have in the meta tags.